safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-dx-prop-expected.txt

CONSOLE MESSAGE: line 36: TypeError: 'undefined' is not an object (evaluating 'dropShadowElement.dx.baseVal = 10')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dx' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL dropShadowElement.dx.baseVal should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'dropShadowElement.dx.baseVal')
